This is the reland of PR https://github.com/pytorch/pytorch/pull/74695, which was reverted due to some internal failures.

It also removes the ShardedTensorInterface change, we will delay that
change later if we found there's a need to do that.

Differential Revision: [D36423976](https://our.internmc.facebook.com/intern/diff/D36423976)

[ghstack-poisoned]

+++ torch/csrc/jit/codegen/cuda/test/test_gpu_validator.h
@@ -371,9 +371,9 @@ inline void testValidate(
 
       TORCH_INTERNAL_ASSERT(
           at_tensor.dim() ==
-              static_cast<int64_t>(TensorDomain::noReductions(
-                                       fusion_input_tv->getMaybeRFactorDomain())
-                                       .size()),
+              TensorDomain::noReductions(
+                  fusion_input_tv->getMaybeRFactorDomain())
+                  .size(),
           "Dimensionality mismatch in inputs.");
     }
   }
@@ -400,10 +400,9 @@ inline void testValidate(
     TORCH_INTERNAL_ASSERT(
         aten_output_tensor.dim() == fusion_output_tensor.dim() &&
             fusion_outputs[j].dim() ==
-                static_cast<int64_t>(
-                    TensorDomain::noReductions(
-                        fusion_output_tv->getMaybeRFactorDomain())
-                        .size()),
+                TensorDomain::noReductions(
+                    fusion_output_tv->getMaybeRFactorDomain())
+                    .size(),
         "Dimensionality mismatch in outputs.");
 
     auto tolerance_values = getTolerance(
